requirements.txt
crewai
openai
langchain
python-dotenv
streamlit
